Report known important platform issues properly
Known important platform issues (the upstream Google Classroom API issue and the Daymap upload issue) are now reported properly to the user through the web interface.
The navigation bar of all error pages now shows the display name of the current user, which is the intended behaviour.
Prior to this commit, the navigation bar of all error pages showed a blank username.