Display timetable using user timezone rather than system timezone
The timetable was previously rendered with the system timezone as an indicator of when days begin, end, etc.
This would be problematic for users of any schools in timezones different to the TaskCollect host.
The timetable is now rendered using the user's school's registered timezone instead.
Small changes have also been made to use Go's time.ParseInLocation function instead of using time.Parse and then filling in the timezone of the returned time.Time