Move GAuthID into a global variable
The in-memory copy of the Google OAuth client ID was previously passed around in the User struct and the (obsolete?) authDB struct.
Considering that the entire TaskCollect codebase only needs access to a single copy of the Google OAuth client ID, it has been isolated into a single GAuthID global variable in package server.